Remove unnecessary WarehousesLogic class
Refactor generate_clients to change the logic for total_weight, delivery_type, and goods assignment.

Move distribute_weight method into DataGenerator class for better encapsulation.
